Changes needed to fix Win64 build warnings.
SVN: r1664@trunk

+++ src/gStructure.h
@@ -103,7 +103,7 @@ namespace mesmer
     //Returns true if atoms have coordinates
     bool ReadStructure();
 
-    int NumAtoms() { return Atoms.size(); }
+    size_t NumAtoms() { return Atoms.size(); }
 
     bool IsAtom() {
       if (Atoms.empty())
